Move KFS grpc service into kfs_frontend
