+++ src/BUILD
@@ -114,10 +114,10 @@ cc_library(
         "grpc_utils.hpp",
         "grpcservermodule.cpp",
         "grpcservermodule.hpp",
+        "kfs_frontend/kfs_grpc_inference_service.cpp",
+        "kfs_frontend/kfs_grpc_inference_service.hpp",
         "kfs_frontend/kfs_utils.cpp",
         "kfs_frontend/kfs_utils.hpp",
-        "kfs_grpc_inference_service.cpp",
-        "kfs_grpc_inference_service.hpp",
         "metric.cpp",
         "metric.hpp",
         "metric_config.cpp",
