+++ src/grpcservermodule.hpp
@@ -20,7 +20,7 @@
 
 #include <grpcpp/server.h>
 
-#include "kfs_grpc_inference_service.hpp"
+#include "kfs_frontend/kfs_grpc_inference_service.hpp"
 #include "model_service.hpp"
 #include "module.hpp"
 #include "prediction_service.hpp"
