+++ src/kfs_frontend/kfs_grpc_inference_service.cpp
@@ -19,28 +19,28 @@
 #include <memory>
 #include <string>
 
-#include "deserialization.hpp"
-#include "execution_context.hpp"
-#include "grpc_utils.hpp"
-#include "kfs_frontend/kfs_utils.hpp"
-#include "metric.hpp"
-#include "modelinstance.hpp"
-#include "modelinstanceunloadguard.hpp"
-#include "modelmanager.hpp"
-#include "ovinferrequestsqueue.hpp"
-#include "pipeline.hpp"
-#include "pipelinedefinition.hpp"
-#include "pipelinedefinitionstatus.hpp"
-#include "pipelinedefinitionunloadguard.hpp"
-#include "prediction_service_utils.hpp"
-#include "serialization.hpp"
-#include "servablemanagermodule.hpp"
-#include "server.hpp"
-#include "status.hpp"
-#include "stringutils.hpp"
-#include "tensorinfo.hpp"
-#include "timer.hpp"
-#include "version.hpp"
+#include "../deserialization.hpp"
+#include "../execution_context.hpp"
+#include "../grpc_utils.hpp"
+#include "../kfs_frontend/kfs_utils.hpp"
+#include "../metric.hpp"
+#include "../modelinstance.hpp"
+#include "../modelinstanceunloadguard.hpp"
+#include "../modelmanager.hpp"
+#include "../ovinferrequestsqueue.hpp"
+#include "../pipeline.hpp"
+#include "../pipelinedefinition.hpp"
+#include "../pipelinedefinitionstatus.hpp"
+#include "../pipelinedefinitionunloadguard.hpp"
+#include "../prediction_service_utils.hpp"
+#include "../serialization.hpp"
+#include "../servablemanagermodule.hpp"
+#include "../server.hpp"
+#include "../status.hpp"
+#include "../stringutils.hpp"
+#include "../tensorinfo.hpp"
+#include "../timer.hpp"
+#include "../version.hpp"
 
 namespace {
 enum : unsigned int {
