+++ src/tensorinfo.cpp
@@ -21,7 +21,7 @@
 #include <string>
 #include <unordered_map>
 
-#include "kfs_grpc_inference_service.hpp"
+#include "kfs_frontend/kfs_grpc_inference_service.hpp"
 #include "logging.hpp"
 
 namespace ovms {
